feat: consoleRunner 생성
